Move seleniumResponse() function to the driver implementation modules (Closes #20)
