+++ lib/driver/RCImpl.js
@@ -95,5 +95,26 @@ module.exports = inherit(CommonImpl, /** @lends RCImpl.prototype */ {
             path: '/selenium-server/driver?cmd=testComplete&sessionId=' + session.getId(),
             method: 'POST'
         }
+    },
+
+    seleniumResponse: function(status, value, sessionId) {
+        status = status || 0;
+
+        var body;
+        if (status === 0) {
+            body = 'OK';
+            if (sessionId || value) {
+                body += ',' + (sessionId || value);
+            }
+        } else {
+            body = 'ERROR: ' + value;
+        }
+
+        return {
+            // TODO: more precise HTTP status based on selenium status code
+            status: status === 0 ? 200 : 500,
+            headers: {},
+            body: [body]
+        };
     }
 });
