+++ lib/driver/WebDriverImpl.js
@@ -4,7 +4,6 @@ var CommonImpl = require('./CommonImpl'),
     inherit = require('inherit'),
     apps = require('../http-apps'),
     normalizeCapabilities = require('../capability-utils').normalizeCapabilities;
-    seleniumResponse = require('./util').seleniumResponse;
 
 var reSessionGet = new RegExp('^/wd/hub/session/[^\/]+$'),
     reSessionCmd = new RegExp('^/wd/hub/session/[^\/]+/.+?$'),
@@ -94,7 +93,7 @@ module.exports = inherit(CommonImpl, /** @lends WebDriverImpl.prototype */ {
     },
 
     getNewSessionResponse: function(session) {
-        return seleniumResponse(0, session.capabilities, session.getId());
+        return this.seleniumResponse(0, session.capabilities, session.getId());
     },
 
     getEndSessionRequest: function(session) {
@@ -102,6 +101,22 @@ module.exports = inherit(CommonImpl, /** @lends WebDriverImpl.prototype */ {
             path: '/wd/hub/session/' + session.getId(),
             method: 'DELETE'
         }
+    },
+
+    seleniumResponse: function(status, value, sessionId) {
+        status = status || 0;
+        var data = {status: status};
+        if (value) {
+            data.value = value;
+        }
+        if (sessionId) {
+            data.sessionId = String(sessionId);
+        }
+        return {
+            // TODO: more precise HTTP status based on selenium status code
+            status: status === 0 ? 200 : 500,
+            headers: {},
+            data: data
+        };
     }
 });
-
