+++ lib/servlets/request.js
@@ -34,7 +34,7 @@ function newSession(req, registry, driver) {
     return registry.getNewSession(req)
         .then(function(session) {
             if (!session) {
-                return q.reject(seleniumResponse(33));
+                return q.reject(driver.seleniumResponse(33));
             }
             // return capabilities received from the node
             return driver.getNewSessionResponse(session);
@@ -44,7 +44,7 @@ function newSession(req, registry, driver) {
 function getSessionInfo(req, registry, driver) {
     return getRequestSession(req, registry, driver)
         .then(function(session) {
-            return seleniumResponse(0, session.capabilities, session.getId());
+            return driver.seleniumResponse(0, session.capabilities, session.getId());
         });
 }
 
@@ -77,20 +77,3 @@ function getRequestSession(req, registry, driver) {
 // 9 UnknownCommand - The requested resource could not be found, or a request was received using an HTTP method that is not supported by the mapped resource
 // 13 UnknownError - An unknown server-side error occurred while processing the command.
 // 33 SessionNotCreatedException - A new session could not be created.
-
-function seleniumResponse(status, value, sessionId) {
-    status = status || 0;
-    var data = {status: status};
-    if (value) {
-        data.value = value;
-    }
-    if (sessionId) {
-        data.sessionId = String(sessionId);
-    }
-    return {
-        // TODO: more precise HTTP status based on selenium status code
-        status: status === 0 ? 200 : 500,
-        headers: {},
-        data: data
-    };
-}
